Change return value to double

+++ include/pros/adi.h
@@ -693,7 +693,7 @@ typedef int32_t adi_potentiometer_t;
 
 adi_potentiometer_t adi_potentiometer_init(uint8_t port, bool new_potentiometer);
 
-int32_t adi_potentiometer_get(adi_potentiometer_t potentiometer);
+double adi_potentiometer_get(adi_potentiometer_t potentiometer);
 
 #ifdef __cplusplus
 }  // namespace c
